Refactor: engine.tsを分離
