+++ src/mobile/engine/manifestAssets/base.json
